🔧 Clean up files after reset to working state
